feat: Add ProductRepository and RepositoryInterface
This commit adds the `ProductRepository` class and the `RepositoryInterface` interface. The `ProductRepository` implements the `RepositoryInterface` and provides methods for creating, updating, deleting, getting, and checking the existence of products. The `RepositoryInterface` defines the common methods that all repositories should implement. These additions will allow for the management of product data in the application.

Note: This commit message does not include any meta information like issue references, tags, or author names.